+++ compiler-core/src/format.rs
@@ -247,8 +247,8 @@ impl<'comments> Formatter<'comments> {
                         .group();
                     ".{".to_doc().append(unqualified).append("}")
                 })
-                .append(if let Some(name) = as_name {
-                    docvec![" as ", name]
+                .append(if let Some(as_name) = as_name {
+                    docvec![" as ", &as_name.name]
                 } else {
                     nil()
                 }),
@@ -1458,6 +1458,10 @@ impl<'comments> Formatter<'comments> {
                 .append(".")
                 .append(label)
                 .to_doc(),
+
+            ClauseGuard::ModuleSelect {
+                module_name, label, ..
+            } => module_name.to_doc().append(".").append(label).to_doc(),
 
             ClauseGuard::Constant(constant) => self.const_expr(constant),
         }