+++ compiler-core/src/type_/environment.rs
@@ -23,6 +23,9 @@ pub struct Environment<'a> {
     pub unused_modules: HashMap<SmolStr, SrcSpan>,
     pub imported_types: HashSet<SmolStr>,
 
+    /// Names of modules that have been imported with as name.
+    pub imported_module_aliases: HashSet<SmolStr>,
+
     /// Values defined in the current function (or the prelude)
     pub scope: im::HashMap<SmolStr, ValueConstructor>,
 
@@ -73,6 +76,7 @@ impl<'a> Environment<'a> {
             scope: prelude.values.clone().into(),
             importable_modules,
             imported_types: HashSet::new(),
+            imported_module_aliases: HashSet::new(),
             current_module,
             warnings,
             entity_usages: vec![HashMap::new()],
@@ -500,8 +504,12 @@ impl<'a> Environment<'a> {
         self.handle_unused(unused);
 
         for (name, location) in self.unused_modules.clone().into_iter() {
-            self.warnings
-                .emit(Warning::UnusedImportedModule { name, location });
+            let warning = if self.imported_module_aliases.contains(&name) {
+                Warning::UnusedImportedModuleAlias { name, location }
+            } else {
+                Warning::UnusedImportedModule { name, location }
+            };
+            self.warnings.emit(warning);
         }
     }