+++ compiler-core/src/type_/tests/warnings.rs
@@ -434,6 +434,42 @@ fn unused_imported_module_with_alias_warnings_test() {
     );
 }
 
+// https://github.com/gleam-lang/gleam/issues/2326
+#[test]
+fn unused_imported_module_with_alias_and_unqualified_name_warnings_test() {
+    let warnings = get_warnings(
+        "import gleam/foo.{bar} as foo",
+        vec![("thepackage", "gleam/foo", "pub fn bar() { 1 }")],
+    );
+    assert!(!warnings.is_empty());
+    assert_eq!(
+        Warning::UnusedImportedValue {
+            name: "bar".into(),
+            location: SrcSpan { start: 18, end: 21 },
+        },
+        warnings[0]
+    );
+    assert_eq!(
+        Warning::UnusedImportedModuleAlias {
+            name: "foo".into(),
+            location: SrcSpan { start: 23, end: 29 },
+        },
+        warnings[1]
+    );
+}
+
+#[test]
+fn unused_imported_module_with_alias_and_unqualified_name_no_warnings_test() {
+    assert_warning!(
+        ("gleam/foo", "pub fn bar() { 1 }"),
+        "import gleam/foo.{bar} as foo\npub fn baz() { bar() }",
+        Warning::UnusedImportedModuleAlias {
+            name: "foo".into(),
+            location: SrcSpan { start: 23, end: 29 },
+        }
+    );
+}
+
 #[test]
 fn unused_imported_module_no_warning_on_used_function_test() {
     assert_no_warnings!(